+++ packages/beta/src/__tests__/api/templateInstances.int.spec.ts
@@ -91,7 +91,7 @@ describe('template instances test', () => {
     expect(result).toEqual(expectedInstances);
   });
 
-  it('should delete template instances', async () => {
+  it.skip('should delete template instances', async () => {
     await client.templates
       .group(externalId)
       .version(1)